src/localstack-docker/pyproject.toml
[project]
name = "localstack-docker"
version = "0.1.0"
description = "LocalStack example with Lambda and S3"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
]
